requirements.txt
crewai
crewai-tools
langchain-google-genai
python-dotenv
google-generativeai
ipython
langchain-groq
